Added contributing guidelines.  Some Renaming.
